backend/requirements.txt
alabaster==0.7.12
amqp==2.3.2
Babel==2.6.0
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-cors-headers==3.0.2
django-des==2.4.1
django-extensions==2.1.4
django-filter==2.0.0
django-rest-auth==0.9.3
django-solo==1.1.3
django-timezone-field==3.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djmail==1.1.0
docutils==0.14
graphviz==0.10.1
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.13.1
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.1.0
mixer==6.1.3
oauthlib==2.1.0
packaging==18.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.1
pygraphviz==1.5
PyJWT==1.7.1
pyparsing==2.3.0
pytest==5.2.2
pytest-django==3.6.0
python-crontab==2.3.5
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
traitlets==4.3.2
urllib3==1.24.1
vine==1.1.4
wcwidth==0.1.7
